requirements.txt
aiohttp==3.9.3
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
cryptography==42.0.5
decorator==5.1.1
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
filelock==3.13.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
identify==2.5.35
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
nodeenv==1.8.0
orjson==3.9.15
outcome==1.3.0.post0
packaging==24.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
pwdlib==0.2.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-dependency==0.6.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.28
SQLAlchemy-Utils==0.41.2
starlette==0.36.3
tornado==6.4
trio==0.25.0
typing_extensions==4.10.0
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
